{"schema":"decodex/commit/1","summary":"Split Decodex app account model modules","authority":"manual"}
